Update README.md
Initialization of Array and Target:

It defines a vector arr containing integer values {2, 3, 4, 10, 40}.

A variable target is set to 10, which is the value we want to search for in the array.



2. Calling the linearSearch Function:

The linearSearch function is called with arr (the array) and target (the value to search for) as arguments. The result of the search (the index of the target) is stored in the variable result.



3. Checking the Search Result:

If result != -1, meaning the target was found, the program prints the index where the target was found (Target found at index).

If result == -1, meaning the target was not found, the program prints Target not found.